refactor: Update audio upload hook and restructure token service imports
