+++ frontend/src/app/(public)/page.tsx
@@ -1,9 +1,8 @@
 'use client';
 
 import { Container, Box, Alert } from '@mui/material';
-
-import { useAudioUpload } from '@/hooks/useAudioUpload';
 import { AudioDropzone, AudioPreview } from '@/components/features/audio';
+import { useAudioUpload } from '@/domain/audio/hook';
 
 export default function HomePage() {
   const {
