+++ frontend/src/domain/audio/hook/index.ts
@@ -1,0 +1,1 @@
+export { useAudioUpload } from './useAudioUpload';
